fix: Fixed a issue typechecking lambdas
Fixed an issue type checking lambda expressions. The resulting type was not being
unified correctly so `map(xs, \(a) -> a.x)` where `a` was a `Foo` would still have the typechecker
believe that `Foo` was of type `A`, which broke mapping over generic arrays of `A`.

Also fixed in this commit was parsing of enum fields and improvements to how enums and records
are represented at codegen (now stored as const functions that construct a new instance).